refactor: move library tests to separate module file
